Workaround broken virtual manifest git dependencies
Do not patch git dependencies to path dependencies and remove the frozen
flag from cargo to be allowed to resolve the git dependencies in the
compile step.
This is a somewhat ugly workaround, as it breaks the reproducibility of
builds to some degree and causes cargo to potentially do network
requests in the compile step, which it should not be allowed to do. For
now it's the best we have as no proper solution is known yet.

Signed-off-by: Julian Scheel <julian@jusst.de>